docker/requirements.txt
feedparser
requests
PyYAML
openai